+++ lib/src/core/utilities/shadows/fx_shadow.dart
@@ -19,43 +19,34 @@ class FxShadow {
       );
 
   ///Gives no shadow
-  static List<BoxShadow> get none => _copyWith(
-        color: Colors.transparent,
-        blurRadius: 0,
-        offset: const Offset(0, 0),
-        spreadRadius: 0,
-      );
+  static List<BoxShadow> get none => _copyWith();
 
   ///Gives custom thin shadow
   static List<BoxShadow> get thin => _copyWith(
-        color: FxColors.blueGray400,
-        blurRadius: 2,
-        offset: const Offset(1, 1),
-        spreadRadius: 1,
+        color: FxColors.dark.withAlpha(10),
+        blurRadius: 6.0,
+        offset: const Offset(0.0, 3.0),
       );
 
   ///Gives custom normal shadow
   static List<BoxShadow> get normal => _copyWith(
-        color: FxColors.blueGray400,
-        blurRadius: 2,
-        offset: const Offset(1, 2),
-        spreadRadius: 1,
+        color: FxColors.dark.withAlpha(20),
+        blurRadius: 6.0,
+        offset: const Offset(0.0, 3.0),
       );
 
   ///Gives custom medium shadow
   static List<BoxShadow> get medium => _copyWith(
-        color: FxColors.blueGray400,
-        blurRadius: 3,
-        offset: const Offset(1, 2),
-        spreadRadius: 2,
+        color: FxColors.dark.withAlpha(40),
+        blurRadius: 6.0,
+        offset: const Offset(0.0, 3.0),
       );
 
   ///Gives custom thik shadow
   static List<BoxShadow> get thik => _copyWith(
-        color: FxColors.blueGray400,
-        blurRadius: 4,
-        offset: const Offset(2, 3),
-        spreadRadius: 2,
+        color: Color(0xFF000000).withAlpha(60),
+        blurRadius: 6.0,
+        offset: const Offset(1.0, 3.0),
       );
 
   static List<BoxShadow> _copyWith({