+++ lib/src/fx_utilities.dart
@@ -4,5 +4,4 @@ export 'core/utilities/radius/fx_border_radius.dart';
 export 'core/utilities/shadows/fx_shadow.dart';
 export 'core/utilities/sizing/fx_height.dart';
 export 'core/utilities/sizing/fx_width.dart';
-export 'core/utilities/spacing/fx_margin.dart';
 export 'core/utilities/spacing/fx_padding.dart';